@@ -8,14 +8,15 @@
   },
   "dependencies": {
     "core-js": "^3.6.5",
-    "vue": "^2.6.11",
-    "vuex": "^3.4.0"
+    "vue": "^3.0.0-beta.1",
+    "vuex": "^4.0.0-alpha.1"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.4.0",
     "@vue/cli-plugin-vuex": "~4.4.0",
     "@vue/cli-service": "~4.4.0",
-    "vue-template-compiler": "^2.6.11"
+    "@vue/compiler-sfc": "^3.0.0-beta.1",
+    "vue-cli-plugin-vue-next": "~0.1.3"
   },
   "browserslist": [
     "> 1%",
